requirements.txt
Django==4.0.5
django-session-timeout==0.0.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
django-filter==21.1
django-webpack-loader
python-dateutil
APScheduler==3.6.1
django-apscheduler==0.3.0
image==1.5.27
django-cors-headers==3.5.0
requests==2.25.0
python-decouple==3.5
aead==0.2
dj_database_url==0.5.0
gunicorn==20.1.0
whitenoise==6.2.0
hubspot-api-client==5.0.1
xlrd==2.0.1
psycopg2==2.8.6
psycopg2-binary==2.8.6